requirements.txt
anthropic==0.7.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
pytest==7.4.3
firecrawl==1.6.8